requirement.txt
python>=3.5.0
tensorflow-gpu>=2.3.0